+++ main-server/src/utils/request_status.test.js
@@ -637,6 +637,104 @@ describe('Test derive request status from request details', () => {
   it('should derive request status correctly', () => {
     const requestStatus = utils.getRequestStatus({
       request_id:
+        'b5974cfbc6015b92e683ab6c4d432b2bc4f7ad56d7eb214b0a4c3e4e36f81e85',
+      min_idp: 1,
+      min_aal: 1,
+      min_ial: 1.1,
+      request_timeout: 86400,
+      idp_id_list: ['idp1', 'idp2'],
+      data_request_list: [
+        {
+          service_id: 'bank_statement',
+          as_id_list: ['as1'],
+          min_as: 1,
+          request_params_hash: 'zaWTDk3EnxboynYfe5P+ygNFyPevkDK1t0TpECOGSvQ=',
+          response_list: [
+            {
+              as_id: 'as1',
+              error_code: 10101,
+            },
+          ],
+        },
+      ],
+      request_message_hash: 'iXVFxq9/i+U1PCxP6KPikpOu2EqA4VTNbE+q5llhW4I=',
+      response_list: [
+        {
+          ial: 2.3,
+          aal: 3,
+          status: 'accept',
+          signature:
+            'VqbDUzBOOwNkhvvFe2oukLJOeTDKOmlJxcyfoUl6BtuEmCqa0QlhCVv7i5HmVzBLwQ5G9MVB0s2NS1zrhI566GYzR+wAhxtBfYN4ADteRsq/nhVlA6VrvQ7LaFSrnls2hyAVKiKaKe1QV97zKwraXzifgJrIdS5Hbv1veqU79pvrIvdAMUeO9IsFS/KGJYnNIg7yr/aURuOHb6fAjBVKEWrk9JL0HGxj0BnTRIyvnD4q21jk6EZTJzarkV0m2VU2I5NI9y+La9gyWRosE2CFsfxBTeSpnWFNZ5B4HLX47cghFmGpxLN1ewpneeFjJF10dh4YWpFHS+0Biudz2FXk/g==',
+          idp_id: 'idp2',
+          valid_ial: null,
+          valid_signature: null,
+        },
+      ],
+      closed: false,
+      timed_out: false,
+      purpose: '',
+      mode: 2,
+      request_type: null,
+      requester_node_id: 'rp1',
+      creation_block_height: 74,
+      creation_chain_id: 'test-chain-NDID',
+    });
+
+    expect(requestStatus).to.equal('errored');
+  });
+
+  it('should derive request status correctly', () => {
+    const requestStatus = utils.getRequestStatus({
+      request_id:
+        'b5974cfbc6015b92e683ab6c4d432b2bc4f7ad56d7eb214b0a4c3e4e36f81e85',
+      min_idp: 1,
+      min_aal: 1,
+      min_ial: 1.1,
+      request_timeout: 86400,
+      idp_id_list: ['idp1', 'idp2'],
+      data_request_list: [
+        {
+          service_id: 'bank_statement',
+          as_id_list: ['as1', 'as2'],
+          min_as: 2,
+          request_params_hash: 'zaWTDk3EnxboynYfe5P+ygNFyPevkDK1t0TpECOGSvQ=',
+          response_list: [
+            {
+              as_id: 'as1',
+              error_code: 10101,
+            },
+          ],
+        },
+      ],
+      request_message_hash: 'iXVFxq9/i+U1PCxP6KPikpOu2EqA4VTNbE+q5llhW4I=',
+      response_list: [
+        {
+          ial: 2.3,
+          aal: 3,
+          status: 'accept',
+          signature:
+            'VqbDUzBOOwNkhvvFe2oukLJOeTDKOmlJxcyfoUl6BtuEmCqa0QlhCVv7i5HmVzBLwQ5G9MVB0s2NS1zrhI566GYzR+wAhxtBfYN4ADteRsq/nhVlA6VrvQ7LaFSrnls2hyAVKiKaKe1QV97zKwraXzifgJrIdS5Hbv1veqU79pvrIvdAMUeO9IsFS/KGJYnNIg7yr/aURuOHb6fAjBVKEWrk9JL0HGxj0BnTRIyvnD4q21jk6EZTJzarkV0m2VU2I5NI9y+La9gyWRosE2CFsfxBTeSpnWFNZ5B4HLX47cghFmGpxLN1ewpneeFjJF10dh4YWpFHS+0Biudz2FXk/g==',
+          idp_id: 'idp2',
+          valid_ial: null,
+          valid_signature: null,
+        },
+      ],
+      closed: false,
+      timed_out: false,
+      purpose: '',
+      mode: 2,
+      request_type: null,
+      requester_node_id: 'rp1',
+      creation_block_height: 74,
+      creation_chain_id: 'test-chain-NDID',
+    });
+
+    expect(requestStatus).to.equal('errored');
+  });
+
+  it('should derive request status correctly', () => {
+    const requestStatus = utils.getRequestStatus({
+      request_id:
         '1804ac0308d22b4203dfc12a930394efa081f30a734c748d2f1966b8c62bb326',
       min_idp: 1,
       min_aal: 3,
@@ -904,6 +1002,197 @@ describe('Test derive request status from request details', () => {
   it('should derive request status correctly', () => {
     const requestStatus = utils.getRequestStatus({
       request_id:
+        '9804ac0308d22b4203dfc12a930394efa081f30a734c748d2f1966b8c62bb326',
+      min_idp: 1,
+      min_aal: 3,
+      min_ial: 2.3,
+      request_timeout: 86400,
+      idp_id_list: ['idp1'],
+      data_request_list: [
+        {
+          service_id: 'bank_statement',
+          as_id_list: ['as1'],
+          min_as: 1,
+          request_params_hash: 'MyUwI7quJS54s6Xh/m4r6WEEa9mL9gKGqPlBrRSoD0A=',
+          response_list: [
+            {
+              as_id: 'as1',
+              error_code: 10101,
+            },
+          ],
+        },
+        {
+          service_id: 'bank_statement_2',
+          as_id_list: ['as1', 'as2'],
+          min_as: 1,
+          request_params_hash: 'MyUwI7quJS54s6Xh/m4r6WEEa9mL9gKGqPlBrRSoD0A=',
+          response_list: [
+            {
+              as_id: 'as1',
+              signed: true,
+              received_data: true,
+            },
+          ],
+        },
+      ],
+      request_message_hash: '/+m2u2JBZ3zADwVdtLiMkvMjONWL8ZsDTqduLp3csGU=',
+      response_list: [
+        {
+          ial: 2.3,
+          aal: 3,
+          status: 'accept',
+          signature:
+            'NHzTRPXXCcCFNHBrA3DSJnVDvXaz+u76Qix55ADLxiWktvJGFY7/wXTK/OCH6Gs1diVM6KNuh6rLP5cPWiaJdkx2zP+3An+1/ox+HqqKb6AKIO0EAKaH3seO1F4wqLrf1ymLN7dobEwntuR3n4XI9eaXsrhNNlTuilYDG4+iA/xzBuCB9QBEhb6hBvBGPsG2BpQiHVveh4t0+6MxdpgkzQuNyzMLqvrj7e71YQjY2cMBC0zGbzcwsyI4rM6NEptsVsf7IJKXW5YNBygp8dxuLCXm5dcksN2MouYp+VU36mTYiAr74sn8isXOsWeVwgbvEOh93uB7ZjQomcU0L1UXDQ==',
+          idp_id: 'idp1',
+          valid_ial: true,
+          valid_signature: true,
+        },
+      ],
+      closed: true,
+      timed_out: false,
+      purpose: '',
+      mode: 2,
+      request_type: null,
+      requester_node_id: 'rp1',
+      creation_block_height: 488,
+      creation_chain_id: 'test-chain-NDID',
+    });
+
+    expect(requestStatus).to.equal('errored');
+  });
+
+  it('should derive request status correctly', () => {
+    const requestStatus = utils.getRequestStatus({
+      request_id:
+        'a804ac0308d22b4203dfc12a930394efa081f30a734c748d2f1966b8c62bb326',
+      min_idp: 1,
+      min_aal: 3,
+      min_ial: 2.3,
+      request_timeout: 86400,
+      idp_id_list: ['idp1'],
+      data_request_list: [
+        {
+          service_id: 'bank_statement',
+          as_id_list: ['as1', 'as2'],
+          min_as: 1,
+          request_params_hash: 'MyUwI7quJS54s6Xh/m4r6WEEa9mL9gKGqPlBrRSoD0A=',
+          response_list: [
+            {
+              as_id: 'as1',
+              error_code: 10101,
+            },
+          ],
+        },
+        {
+          service_id: 'bank_statement_2',
+          as_id_list: ['as1', 'as2'],
+          min_as: 1,
+          request_params_hash: 'MyUwI7quJS54s6Xh/m4r6WEEa9mL9gKGqPlBrRSoD0A=',
+          response_list: [
+            {
+              as_id: 'as1',
+              signed: true,
+              received_data: true,
+            },
+          ],
+        },
+      ],
+      request_message_hash: '/+m2u2JBZ3zADwVdtLiMkvMjONWL8ZsDTqduLp3csGU=',
+      response_list: [
+        {
+          ial: 2.3,
+          aal: 3,
+          status: 'accept',
+          signature:
+            'NHzTRPXXCcCFNHBrA3DSJnVDvXaz+u76Qix55ADLxiWktvJGFY7/wXTK/OCH6Gs1diVM6KNuh6rLP5cPWiaJdkx2zP+3An+1/ox+HqqKb6AKIO0EAKaH3seO1F4wqLrf1ymLN7dobEwntuR3n4XI9eaXsrhNNlTuilYDG4+iA/xzBuCB9QBEhb6hBvBGPsG2BpQiHVveh4t0+6MxdpgkzQuNyzMLqvrj7e71YQjY2cMBC0zGbzcwsyI4rM6NEptsVsf7IJKXW5YNBygp8dxuLCXm5dcksN2MouYp+VU36mTYiAr74sn8isXOsWeVwgbvEOh93uB7ZjQomcU0L1UXDQ==',
+          idp_id: 'idp1',
+          valid_ial: true,
+          valid_signature: true,
+        },
+      ],
+      closed: true,
+      timed_out: false,
+      purpose: '',
+      mode: 2,
+      request_type: null,
+      requester_node_id: 'rp1',
+      creation_block_height: 488,
+      creation_chain_id: 'test-chain-NDID',
+    });
+
+    expect(requestStatus).to.equal('confirmed');
+  });
+
+  it('should derive request status correctly', () => {
+    const requestStatus = utils.getRequestStatus({
+      request_id:
+        'a804ac0308d22b4203dfc12a930394efa081f30a734c748d2f1966b8c62bb326',
+      min_idp: 1,
+      min_aal: 3,
+      min_ial: 2.3,
+      request_timeout: 86400,
+      idp_id_list: ['idp1'],
+      data_request_list: [
+        {
+          service_id: 'bank_statement',
+          as_id_list: ['as1', 'as2'],
+          min_as: 1,
+          request_params_hash: 'MyUwI7quJS54s6Xh/m4r6WEEa9mL9gKGqPlBrRSoD0A=',
+          response_list: [
+            {
+              as_id: 'as1',
+              error_code: 10101,
+            },
+            {
+              as_id: 'as2',
+              signed: true,
+              received_data: true,
+            },
+          ],
+        },
+        {
+          service_id: 'bank_statement_2',
+          as_id_list: ['as1', 'as2'],
+          min_as: 1,
+          request_params_hash: 'MyUwI7quJS54s6Xh/m4r6WEEa9mL9gKGqPlBrRSoD0A=',
+          response_list: [
+            {
+              as_id: 'as1',
+              signed: true,
+              received_data: true,
+            },
+          ],
+        },
+      ],
+      request_message_hash: '/+m2u2JBZ3zADwVdtLiMkvMjONWL8ZsDTqduLp3csGU=',
+      response_list: [
+        {
+          ial: 2.3,
+          aal: 3,
+          status: 'accept',
+          signature:
+            'NHzTRPXXCcCFNHBrA3DSJnVDvXaz+u76Qix55ADLxiWktvJGFY7/wXTK/OCH6Gs1diVM6KNuh6rLP5cPWiaJdkx2zP+3An+1/ox+HqqKb6AKIO0EAKaH3seO1F4wqLrf1ymLN7dobEwntuR3n4XI9eaXsrhNNlTuilYDG4+iA/xzBuCB9QBEhb6hBvBGPsG2BpQiHVveh4t0+6MxdpgkzQuNyzMLqvrj7e71YQjY2cMBC0zGbzcwsyI4rM6NEptsVsf7IJKXW5YNBygp8dxuLCXm5dcksN2MouYp+VU36mTYiAr74sn8isXOsWeVwgbvEOh93uB7ZjQomcU0L1UXDQ==',
+          idp_id: 'idp1',
+          valid_ial: true,
+          valid_signature: true,
+        },
+      ],
+      closed: true,
+      timed_out: false,
+      purpose: '',
+      mode: 2,
+      request_type: null,
+      requester_node_id: 'rp1',
+      creation_block_height: 488,
+      creation_chain_id: 'test-chain-NDID',
+    });
+
+    expect(requestStatus).to.equal('completed');
+  });
+
+  it('should derive request status correctly', () => {
+    const requestStatus = utils.getRequestStatus({
+      request_id:
         '1804ac0308d22b4203dfc12a930394efa081f30a734c748d2f1966b8c62bb326',
       min_idp: 1,
       min_aal: 3,
@@ -960,7 +1249,7 @@ describe('Test derive request status from request details', () => {
       creation_chain_id: 'test-chain-NDID',
     });
 
-    expect(requestStatus).to.equal('confirmed');
+    expect(requestStatus).to.equal('partial_completed');
   });
 
   it('should derive request status correctly', () => {
@@ -1027,7 +1316,7 @@ describe('Test derive request status from request details', () => {
       creation_chain_id: 'test-chain-NDID',
     });
 
-    expect(requestStatus).to.equal('confirmed');
+    expect(requestStatus).to.equal('partial_completed');
   });
 
   it('should derive request status correctly', () => {
@@ -1160,7 +1449,7 @@ describe('Test derive request status from request details', () => {
       creation_chain_id: 'test-chain-NDID',
     });
 
-    expect(requestStatus).to.equal('errored');
+    expect(requestStatus).to.equal('partial_completed');
   });
 
   it('should derive request status correctly', () => {
@@ -1226,7 +1515,7 @@ describe('Test derive request status from request details', () => {
       creation_chain_id: 'test-chain-NDID',
     });
 
-    expect(requestStatus).to.equal('errored');
+    expect(requestStatus).to.equal('partial_completed');
   });
 
   it('should derive request status correctly', () => {